refactor: split App.tsx into focused hooks and PanelHost component (#300)
Extract three custom hooks (useTerminalSettings, useAppTheme, useShellData)
and a PanelHost component to reduce App.tsx from 2203 to 1128 lines (49%).
