+++ src/hooks/useShellData.ts
@@ -1,0 +1,96 @@
+import { useState, useEffect, useCallback, useMemo } from "react";
+import { invoke } from "@tauri-apps/api/core";
+import { useUiStore } from "../stores/uiStore";
+import type { PanelKey } from "./usePanels";
+
+interface ProjectInfo {
+  id: number;
+  name: string;
+  local_path: string;
+  framework: string | null;
+}
+
+interface WorktreeInfo {
+  id: number;
+  project_id: number;
+  branch_name: string;
+  path: string;
+  status: string;
+}
+
+export function useShellData(openPanel: (name: PanelKey) => void) {
+  const {
+    selectedProjectId,
+    selectedWorktreeId,
+    setSelectedProjectId,
+    setSelectedWorktreeId,
+    setSelectedWorktreePath,
+    setSelectedWorktreeName,
+    setShowSettings,
+  } = useUiStore();
+
+  const [allProjects, setAllProjects] = useState<ProjectInfo[]>([]);
+  const [allWorktrees, setAllWorktrees] = useState<
+    Array<WorktreeInfo & { projectName: string }>
+  >([]);
+
+  useEffect(() => {
+    async function loadSwitcherData() {
+      try {
+        const projects = await invoke<ProjectInfo[]>("list_projects");
+        setAllProjects(projects);
+
+        if (
+          projects.length === 0 &&
+          !localStorage.getItem("workroot:onboarded")
+        ) {
+          openPanel("onboarding");
+        }
+
+        const wtResults = await Promise.all(
+          projects.map(async (p) => {
+            try {
+              const wts = await invoke<WorktreeInfo[]>(
+                "list_project_worktrees",
+                { projectId: p.id },
+              );
+              return wts.map((wt) => ({ ...wt, projectName: p.name }));
+            } catch {
+              return [];
+            }
+          }),
+        );
+        setAllWorktrees(wtResults.flat());
+      } catch {
+        // ignore — commands just won't have project/worktree entries
+      }
+    }
+    loadSwitcherData();
+  }, [selectedProjectId, selectedWorktreeId, openPanel]);
+
+  const selectedProjectName = useMemo(
+    () => allProjects.find((p) => p.id === selectedProjectId)?.name ?? null,
+    [allProjects, selectedProjectId],
+  );
+
+  const selectWorktree = useCallback(
+    (wt: WorktreeInfo & { projectName: string }) => {
+      setSelectedProjectId(wt.project_id);
+      setSelectedWorktreeId(wt.id);
+      setSelectedWorktreePath(wt.path);
+      setSelectedWorktreeName(wt.branch_name);
+      setShowSettings(false);
+    },
+    [
+      setSelectedProjectId,
+      setSelectedWorktreeId,
+      setSelectedWorktreePath,
+      setSelectedWorktreeName,
+      setShowSettings,
+    ],
+  );
+
+  return { allProjects, allWorktrees, selectedProjectName, selectWorktree };
+}
+
+export type { ProjectInfo, WorktreeInfo };
